backend/requirements.txt
# HuntPlan AI — Backend Dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.35
alembic>=1.13.2
asyncpg>=0.29.0
geoalchemy2>=0.15.2
pgvector>=0.3.5

# HTTP / Scraping
httpx>=0.27.2
beautifulsoup4>=4.12.3
lxml>=5.3.0

# PDF parsing / OCR
pdfplumber>=0.11.0

# AI / LLM
anthropic>=0.34.0

# Geospatial
shapely>=2.0.6
pyproj>=3.7.0
geojson>=3.1.0

# Background tasks
celery[redis]>=5.4.0

# Utilities
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
orjson>=3.10.7

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
